refactor(stores): move session subpackage (#1960 slice 1)
