+++ tests/core/test_turn_store.py
@@ -10,7 +10,7 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from factory.infrastructure.stores.turn_store import TurnStore
+from factory.infrastructure.stores.session.turn_store import TurnStore
 
 
 @pytest.fixture
@@ -429,7 +429,7 @@ class TestPoolSessions:
         )
         await db.commit()
 
-        from factory.infrastructure.stores.turn_store_queries import backfill_sessions
+        from factory.infrastructure.stores.session.turn_store_queries import backfill_sessions
 
         await backfill_sessions(db)
         await backfill_sessions(db)  # second call must be a no-op
