+++ tests/integration/test_crash_recovery.py
@@ -29,7 +29,7 @@ from uuid import UUID, uuid4
 
 import pytest
 
-from factory.infrastructure.stores.turn_store import TurnStore
+from factory.infrastructure.stores.session.turn_store import TurnStore
 from factory.infrastructure.turn_writer.writer import TurnWriter
 from roxabi_contracts.envelope import CONTRACT_VERSION, new_job_id
 from roxabi_contracts.turns import (
